factor out controller, add region tags.
